typechecker: Reuse variable for bindings/destructuring
Removes the burden of creating a new variable from the compiler and
avoids extra copies in the generated C++ code.

Removes the `CheckedStatement::DestructuringAssignment` variant, since
now the typechecker pushes the declaration statements directly into the
current block and buffers the last one so that it can return it. None of
those statements are yield statements, so they don't need to be checked
by the usual .typecheck_block <-> .typecheck_statement relationship.