codegen: Only `#include` modules when their C++ is needed
There are some instances where only the implementation files for modules
need to include types from other modules, because the APIs of the
modules themselves don't include said types. Here we make sure that
module dependencies start with an empty set so that we include modules
only when the C++ that we emit mentions the types or functions that are
exported.

Also, in `selfhost/git.jakt`, we use a function from `utility` only at
compile time, and nowhere at runtime. This means that the generated C++
doesn't need to include "utility.h" at all!

This reduces include overhead, and with this accidental recompilations
due to changing headers.